+++ lib/playwright_ex/channels/browser_context.ex
@@ -223,7 +223,7 @@ defmodule PlaywrightEx.BrowserContext do
     NimbleOptions.new!(
       connection: PlaywrightEx.Channel.connection_opt(),
       timeout: PlaywrightEx.Channel.timeout_opt(),
-      content: [
+      source: [
         type: :string,
         required: true,
         doc: "Raw JavaScript code to be evaluated in all pages before any scripts run."